change height to an integer and minor updates to the UI
- The value of the height is now being saved as an Integer value. The UI now shows the measurement type of the values. The keyboard is different and some minor quality of life changes.

+++ fittrackapp/src/main/java/com/berbas/fittrackapp/screens/bluetooth/BluetoothSyncViewModel.kt
@@ -170,7 +170,7 @@ class BluetoothSyncViewModel @Inject constructor(
             lastname = personParts[1].substringAfter("="),
             birthday = personParts[2].substringAfter("="),
             gender = personParts[3].substringAfter("="),
-            height = personParts[4].substringAfter("=").toDouble(),
+            height = personParts[4].substringAfter("=").toInt(),
             weight = personParts[5].substringAfter("=").toDouble()
         )
     }
