+++ fittrackapp/src/main/java/com/berbas/fittrackapp/screens/profile/PersonEvent.kt
@@ -7,6 +7,6 @@ sealed interface PersonEvent {
     data class SetGender(val gender: String): PersonEvent
     data class SetBirthday(val birthday: String): PersonEvent
     data class SetWeight(val weight: Double): PersonEvent
-    data class SetHeight(val height: Double): PersonEvent
+    data class SetHeight(val height: Int): PersonEvent
 
 }
